Create React Components Tree
