+++ src/components/create/FilterOverlay.jsx
@@ -1,0 +1,60 @@
+import plusIcon from '../../assets/icons/plus-icon.png';
+
+function FilterOverlay() {
+  return (
+    <div className="add-item__filter-overlay hide">
+      <ul className="add-item__filter-list">
+        <li className="add-item__filter-item">
+          <div className="add-item__filter-content">
+            <img className="add-item__filter-icon" src={plusIcon} alt="Item" />
+            Egg
+          </div>
+        </li>
+        <li className="add-item__filter-item">
+          <div className="add-item__filter-content">
+            <img className="add-item__filter-icon" src={plusIcon} alt="Item" />
+            Egg
+          </div>
+        </li>
+        <li className="add-item__filter-item">
+          <div className="add-item__filter-content">
+            <img className="add-item__filter-icon" src={plusIcon} alt="Item" />
+            Egg
+          </div>
+        </li>
+        <li className="add-item__filter-item">
+          <div className="add-item__filter-content">
+            <img className="add-item__filter-icon" src={plusIcon} alt="Item" />
+            Egg
+          </div>
+        </li>
+        <li className="add-item__filter-item">
+          <div className="add-item__filter-content">
+            <img className="add-item__filter-icon" src={plusIcon} alt="Item" />
+            Egg
+          </div>
+        </li>
+        <li className="add-item__filter-item">
+          <div className="add-item__filter-content">
+            <img className="add-item__filter-icon" src={plusIcon} alt="Item" />
+            Egg
+          </div>
+        </li>
+        <li className="add-item__filter-item">
+          <div className="add-item__filter-content">
+            <img className="add-item__filter-icon" src={plusIcon} alt="Item" />
+            Egg
+          </div>
+        </li>
+        <li className="add-item__filter-item">
+          <div className="add-item__filter-content">
+            <img className="add-item__filter-icon" src={plusIcon} alt="Item" />
+            Egg
+          </div>
+        </li>
+      </ul>
+    </div>
+  );
+}
+
+export default FilterOverlay;
